+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@temporalio/ui",
-  "version": "2.1.11",
+  "version": "2.1.12",
   "type": "module",
   "description": "Temporal.io UI",
   "keywords": [
