+++ src/lib/components/event/event-details-row.svelte
@@ -13,7 +13,7 @@
     getCodeBlockValue,
   } from '$lib/utilities/get-single-attribute-for-event';
 
-  import CodeBlock from '../code-block.svelte';
+  import CodeBlock from '$lib/holocene/code-block.svelte';
   import Link from '../link.svelte';
   import Copyable from '../copyable.svelte';
