+++ src/lib/pages/workflow-pending-activities.svelte
@@ -6,7 +6,7 @@
   import Badge from '$holocene/badge.svelte';
   import EmptyState from '$lib/components/empty-state.svelte';
   import Link from '$lib/components/link.svelte';
-  import CodeBlock from '$lib/components/code-block.svelte';
+  import CodeBlock from '$lib/holocene/code-block.svelte';
   import {
     formatDate,
     getDuration,
