+++ src/lib/pages/workflow-query.svelte
@@ -3,7 +3,7 @@
   import { getQuery, getQueryTypes } from '$lib/services/query-service';
   import { workflowRun } from '$lib/stores/workflow-run';
 
-  import CodeBlock from '$lib/components/code-block.svelte';
+  import CodeBlock from '$lib/holocene/code-block.svelte';
   import Option from '$lib/components/select/option.svelte';
   import EmptyState from '$lib/components/empty-state.svelte';
   import Select from '$lib/components/select/select.svelte';
