+++ src/routes/fiction/chapters/accordion.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import Chapter from '../_chapter.svelte';
   import Accordion from '$holocene/accordion.svelte';
-  import CodeBlock from '$lib/components/code-block.svelte';
+  import CodeBlock from '$lib/holocene/code-block.svelte';
   import Table from '$holocene/table/table.svelte';
   import TableHeaderRow from '$lib/holocene/table/table-header-row.svelte';
   import TableRow from '$lib/holocene/table/table-row.svelte';
