+++ src/routes/namespaces/[namespace]/archival/index@root.svelte
@@ -61,7 +61,7 @@
   import WorkflowFilters from './_workflow-filters.svelte';
   import Pagination from '$holocene/pagination.svelte';
   import EmptyState from '$lib/components/empty-state.svelte';
-  import CodeBlock from '$lib/components/code-block.svelte';
+  import CodeBlock from '$lib/holocene/code-block.svelte';
   import { timeFormat } from '$lib/stores/time-format';
   import PageTitle from '$lib/holocene/page-title.svelte';
   import { page } from '$app/stores';
